requirements.txt
streamlit
pandas
numpy
altair[all]
scikit-learn
xgboost
joblib
imblearn
google-cloud-language
google-cloud-storage
pydantic
cloudpickle